+++ src/interface_adapter/CompositeViewModel.java
@@ -46,7 +46,6 @@ public class CompositeViewModel extends JPanel implements PropertyChangeListener
     public JPanel gridPanel;
 
     public CompositeViewModel() {
-        setLayout(new GridLayout(3, 2)); // Adjust layout as needed;
 
         this.relatedArtistsViewModel = new RelatedArtistsViewModel();
         this.topGenreViewModel = new TopGenreViewModel();
@@ -54,55 +53,6 @@ public class CompositeViewModel extends JPanel implements PropertyChangeListener
         this.topAlbumViewModel = new TopAlbumViewModel();
         this.topArtistsViewModel = new TopArtistsViewModel();
         this.getValenceViewModel = new GetValenceViewModel();
-        LoggedInViewModel loggedInViewModel = new LoggedInViewModel();
-
-        // Create Back button
-        JButton backButton = new JButton("Back");
-        backButton.addActionListener(new ActionListener() {
-            @Override
-            public void actionPerformed(ActionEvent evt) {
-                String actionCommand = evt.getActionCommand();
-                System.out.println(("Click" + evt.getActionCommand()));
-                Container parent = getParent();
-                JPanel selectedView = null;
-                // Get the parent container (assuming LoggedInView)
-                ViewManagerModel viewManagerModel = new ViewManagerModel();
-                LoggedInView loggedInView = LoggedInUseCaseFactory.create(viewManagerModel, loggedInViewModel, CompositeViewModel.this);
-                if (loggedInView != null) {
-                    selectedView = loggedInView;
-                    showLoggedInView(loggedInView); // Show LoggedInView again
-                }
-            }
-
-            public void showLoggedInView(JPanel loggedInView) {
-                // Clear the existing content of the LoggedInView and show the selected use case view
-                // Here's what the showUseCaseView method does:
-                //
-                //It Clears Existing Content: Before displaying a new view, it removes any existing components from the LoggedInView. This ensures that only one view is visible at a time.
-                //
-                //It Adds the Selected Use Case View: It adds the specified use case view (useCase1View, useCase2View, or useCase3View) to the LoggedInView.
-                //
-                //It Refreshes the UI: After updating the contents, it calls revalidate() to inform Swing to re-layout the components and repaint() to refresh the UI, ensuring that the changes made are reflected visually.
-                removeAll();
-                add(loggedInView);
-                revalidate();
-                repaint();
-            }
-        });
-        // Use a separate JPanel for grid layout to organize view models
-        gridPanel = new JPanel(new GridLayout(3, 2)); // Adjust layout as needed
-
-        // Add view panels to the gridPanel
-        gridPanel.add(getValenceViewModel.getViewPanel());
-        gridPanel.add(relatedArtistsViewModel.getViewPanel());
-        gridPanel.add(topGenreViewModel.getViewPanel());
-        gridPanel.add(topSongsViewModel.getViewPanel());
-        gridPanel.add(topAlbumViewModel.getViewPanel());
-        gridPanel.add(topArtistsViewModel.getViewPanel());
-
-        add(backButton, BorderLayout.SOUTH); // Add the Back button at the top
-        add(gridPanel, BorderLayout.CENTER); // Add the gridPanel to hold view models
-
 
     }
 
@@ -137,8 +87,8 @@ public class CompositeViewModel extends JPanel implements PropertyChangeListener
     }
 
     public JPanel getGridPanel() {
-        gridPanel = new JPanel(new GridLayout(3, 2)); // Adjust layout as needed
-
+        JPanel result = new JPanel();
+        gridPanel = new JPanel(new GridLayout(2, 3)); // Adjust layout as needed
         // Add view panels to the gridPanel
         gridPanel.add(getValenceViewModel.getViewPanel());
         gridPanel.add(relatedArtistsViewModel.getViewPanel());
@@ -147,43 +97,8 @@ public class CompositeViewModel extends JPanel implements PropertyChangeListener
         gridPanel.add(topAlbumViewModel.getViewPanel());
         gridPanel.add(topArtistsViewModel.getViewPanel());
 
-        JButton backButton = new JButton("Back");
-        backButton.addActionListener(new ActionListener() {
-            @Override
-            public void actionPerformed(ActionEvent evt) {
-                String actionCommand = evt.getActionCommand();
-                System.out.println(("Click" + evt.getActionCommand()));
-                Container parent = getParent();
-                JPanel selectedView = null;
-                // Get the parent container (assuming LoggedInView)
-                ViewManagerModel viewManagerModel = new ViewManagerModel();
-                LoggedInViewModel loggedInViewModel = new LoggedInViewModel();
-                LoggedInView loggedInView = LoggedInUseCaseFactory.create(viewManagerModel, loggedInViewModel, CompositeViewModel.this);
-                if (loggedInView != null) {
-                    selectedView = loggedInView;
-                    showLoggedInView(loggedInView); // Show LoggedInView again
-                }
-            }
+        result.add(gridPanel, BorderLayout.CENTER); // Add the gridPanel to hold view models
 
-            public void showLoggedInView(JPanel loggedInView) {
-                // Clear the existing content of the LoggedInView and show the selected use case view
-                // Here's what the showUseCaseView method does:
-                //
-                //It Clears Existing Content: Before displaying a new view, it removes any existing components from the LoggedInView. This ensures that only one view is visible at a time.
-                //
-                //It Adds the Selected Use Case View: It adds the specified use case view (useCase1View, useCase2View, or useCase3View) to the LoggedInView.
-                //
-                //It Refreshes the UI: After updating the contents, it calls revalidate() to inform Swing to re-layout the components and repaint() to refresh the UI, ensuring that the changes made are reflected visually.
-                removeAll();
-                add(loggedInView);
-                revalidate();
-                repaint();
-            }
-        });
-
-        gridPanel.add(backButton, BorderLayout.SOUTH); // Add the Back button at the top
-        add(gridPanel, BorderLayout.CENTER); // Add the gridPanel to hold view models
-
-        return gridPanel;
+        return result;
     }
 }
